CodeGen from PR 11908 in Azure/azure-rest-api-specs Merge 6de645020f9f51b6c602cdce248c7e06f397120e into 82dc61d402278f8515c378fbf87ea468d6a1c18d

+++ sdk/alertsmanagement/azure-mgmt-alertsmanagement/azure/mgmt/alertsmanagement/operations/_alerts_operations.py
@@ -147,7 +147,7 @@ class AlertsOperations(object):
          response. Permissible value is between 1 to 250. When the
          "includeContent"  filter is selected, maximum value allowed is 25.
          Default value is 25.
-        :type page_count: int
+        :type page_count: long
         :param sort_by: Sort the query results by input field,  Default value
          is 'lastModifiedDateTime'. Possible values include: 'name',
          'severity', 'alertState', 'monitorCondition', 'targetResource',
@@ -220,7 +220,7 @@ class AlertsOperations(object):
                 if include_egress_config is not None:
                     query_parameters['includeEgressConfig'] = self._serialize.query("include_egress_config", include_egress_config, 'bool')
                 if page_count is not None:
-                    query_parameters['pageCount'] = self._serialize.query("page_count", page_count, 'int')
+                    query_parameters['pageCount'] = self._serialize.query("page_count", page_count, 'long')
                 if sort_by is not None:
                     query_parameters['sortBy'] = self._serialize.query("sort_by", sort_by, 'str')
                 if sort_order is not None:
